pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
# See https://www.python.org/dev/peps/pep-0621/
name = "aiida-quad"
dynamic = ["version"]  # read from aiida_quad/src/__init__.py
description = "This is the template for an AiiDA plugin"
authors = [{name = "QMat", email = "hung.dangthe@phenikaa-uni.edu.vn"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Development Status :: 3 - Alpha",
    "Framework :: AiiDA"
]
keywords = ["aiida", "plugin"]
requires-python = ">=3.9"
dependencies = [
    "aiida-core>=2.5,<3",
    "voluptuous"
]

[project.optional-dependencies]
pre-commit = [
  'pre-commit~=3.5',
]
docs = [
    "sphinx",
    "sphinxcontrib-contentui",
    "sphinxcontrib-details-directive",
    "furo",
    "markupsafe<2.1"
]

[project.urls]
Source = "https://github.com/qmat-group/aiida-quad"

[project.entry-points."aiida.data"]
"quad" = "aiida_quad.data:DiffParameters"

[project.entry-points."aiida.calculations"]
"quad" = "aiida_quad.calculations:DiffCalculation"

[project.entry-points."aiida.parsers"]
"quad" = "aiida_quad.parsers:DiffParser"

[project.entry-points."aiida.cmdline.data"]
"quad" = "aiida_quad.cli:data_cli"

[tool.pytest.ini_options]
# Configuration for [pytest](https://docs.pytest.org)
python_files = "test_*.py example_*.py"
addopts = "--pdbcls=IPython.terminal.debugger:TerminalPdb"
filterwarnings = [
    "ignore::DeprecationWarning:aiida:",
    "ignore:Creating AiiDA configuration folder:",
    "ignore::DeprecationWarning:plumpy:",
    "ignore::DeprecationWarning:yaml:",
]


[tool.coverage.run]
# Configuration of [coverage.py](https://coverage.readthedocs.io)
# reporting which lines of your plugin are covered by tests
source = ["src/aiida_quad"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
ignore = [
  'F403',  # Star imports unable to detect undefined names
  'F405',  # Import may be undefined or defined from star imports
  'PLR0911',  # Too many return statements
  'PLR0912',  # Too many branches
  'PLR0913',  # Too many arguments in function definition
  'PLR0915',  # Too many statements
  'PLR2004',  # Magic value used in comparison
  'RUF005',  # Consider iterable unpacking instead of concatenation
  'RUF012'  # Mutable class attributes should be annotated with `typing.ClassVar`
]
select = [
  'E',  # pydocstyle
  'W',  # pydocstyle
  'F',  # pyflakes
  'I',  # isort
  'N',  # pep8-naming
  'PLC',  # pylint-convention
  'PLE',  # pylint-error
  'PLR',  # pylint-refactor
  'PLW',  # pylint-warning
  'RUF'  # ruff
]

## Hatch configurations

[tool.hatch.version]
path = "src/aiida_quad/__init__.py"

[tool.hatch.envs.hatch-test]
dependencies = [
    'pgtest~=1.3,>=1.3.1',
    'coverage~=7.0',
    'pytest~=7.0',
    "pytest-cov~=4.1",
    "ipdb"
]

[tool.hatch.envs.hatch-test.scripts]
# These are the efault scripts provided by hatch.
# The have been copied to make the execution more transparent

# This command is run with the command `hatch test`
run = "pytest{env:HATCH_TEST_ARGS:} {args}"
# The three commands below are run with the command `hatch test --coverage`
run-cov = "coverage run -m pytest{env:HATCH_TEST_ARGS:} {args}"
cov-combine = "coverage combine"
cov-report = "coverage report"

[[tool.hatch.envs.hatch-test.matrix]]
python = ["3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.hatch-static-analysis]
dependencies = ["ruff==0.4.3"]

[tool.hatch.envs.hatch-static-analysis.scripts]
# Fixes are executed with `hatch fmt`.
# Checks are executed with `hatch fmt --check`.

format-check = "ruff format --check --config pyproject.toml {args:.}"
format-fix = "ruff format --config pyproject.toml {args:.}"
lint-check = "ruff check --config pyproject.toml {args:.}"
lint-fix = "ruff check --config pyproject.toml --fix --exit-non-zero-on-fix --show-fixes {args:.}"

[tool.hatch.envs.docs]
features = ["docs"]

[tool.hatch.envs.docs.scripts]
build = [
  "make -C docs"
]
